Compare Faces with multi-faces
compare function now has the capabillities to run through source images to check target image agansit it. If the user is found the program will display "Face match found with <image_name>". Otherwise the program will display "Face doesn't match with <image_name>". If there are multiple faces in the target image the program will still know whether there is or is not a match with the current source image.